+++ resources/js/vue/index.ts
@@ -1,2 +1,3 @@
 export { default as Ga4Snippet } from './Ga4Snippet.vue'
 export { default as GaOverview } from './GaOverview.vue'
+export { default as GaTopContent } from './GaTopContent.vue'